Add files via upload
This script can be used as a stop gap if you are not ready to implement security defaults or conditional access. It will disable the legacy email protocols on the user level. You can confirm with two methods that this was successful. First, the line in the script will validate this and second you can log into the Microsoft admin center and verify on the user flyout menu.